chore(graph): move desktop suite graph related stuff back to suite
