+++ packages/suite/src/components/suite/TransactionsGraph/components/CustomTooltipDashboard.tsx
@@ -3,10 +3,10 @@ import React from 'react';
 import { TooltipProps } from 'recharts';
 
 import { useFormatters } from '@suite-common/formatters';
-import { CommonAggregatedHistory } from '@suite-common/wallet-graph';
 
 import { Props as GraphProps, FiatGraphProps } from '../definitions';
 import { CustomTooltipBase } from './CustomTooltipBase';
+import { CommonAggregatedHistory } from '../../../../types/wallet/graph';
 
 interface CustomTooltipDashboardProps extends TooltipProps<number, any> {
     selectedRange: GraphProps['selectedRange'];
