+++ packages/suite/src/views/dashboard/components/PortfolioCard/components/DashboardGraph.tsx
@@ -9,9 +9,10 @@ import * as graphActions from '@wallet-actions/graphActions';
 import { useActions, useSelector } from '@suite-hooks';
 import { Account } from '@wallet-types';
 
-import { getMinMaxValueFromData, AggregatedDashboardHistory } from '@suite-common/wallet-graph';
 import { variables, Button } from '@trezor/components';
 import { calcTicks, calcTicksFromData } from '@suite-common/suite-utils';
+import { AggregatedDashboardHistory } from '../../../../../types/wallet/graph';
+import { getMinMaxValueFromData } from '../../../../../utils/wallet/graphUtils';
 
 const Wrapper = styled.div`
     display: flex;
