+++ packages/suite/tsconfig.json
@@ -54,9 +54,6 @@
             "path": "../../suite-common/wallet-core"
         },
         {
-            "path": "../../suite-common/wallet-graph"
-        },
-        {
             "path": "../../suite-common/wallet-types"
         },
         {
