+++ suite-common/graph-neue/package.json
@@ -21,7 +21,7 @@
         "@suite-common/wallet-types": "workspace:*",
         "@suite-common/wallet-utils": "workspace:*",
         "@trezor/blockchain-link": "workspace:*",
-        "@trezor/connect": "9.0.4",
+        "@trezor/connect": "workspace:9.0.4",
         "bignumber.js": "^9.1.0",
         "date-fns": "^2.29.1",
         "react": "18.2.0",
