+++ suite-common/graph-neue/tsconfig.json
@@ -1,5 +1,18 @@
 {
     "extends": "../../tsconfig.json",
     "compilerOptions": { "outDir": "libDev" },
-    "references": []
+    "references": [
+        { "path": "../fiat-services" },
+        { "path": "../suite-config" },
+        { "path": "../suite-utils" },
+        { "path": "../test-utils" },
+        { "path": "../wallet-config" },
+        { "path": "../wallet-core" },
+        { "path": "../wallet-types" },
+        { "path": "../wallet-utils" },
+        {
+            "path": "../../packages/blockchain-link"
+        },
+        { "path": "../../packages/connect" }
+    ]
 }
