+++ suite-common/test-utils/src/extraDependenciesMock.ts
@@ -9,7 +9,7 @@ const mockedConsoleLog = (...args: any) => {
     // we don't want to see console.log in tests because it's too noisy
     if (process.env.NODE_ENV !== 'test') {
         // eslint-disable-next-line no-console
-        // console.log(...args);
+        console.log(...args);
     }
 };
