--- suite-common/wallet-graph/README.md
@@ -1,5 +1,0 @@
-# @suite-common/wallet-graph
-
-This package contains all graph related stuff.
-
-TODO: When both graphs (in Suite desktop and the mobile app) will be unified to represent the same data, move here the rest of graph functionality (actions, reducer, middleware) from `@trezor/suite`.
