--- suite-common/wallet-graph/jest.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-    preset: '../../jest.config.base.js',
-};
