--- suite-common/wallet-graph/package.json
@@ -1,34 +1,0 @@
-{
-    "name": "@suite-common/wallet-graph",
-    "version": "1.0.0",
-    "private": true,
-    "license": "See LICENSE.md in repo root",
-    "sideEffects": false,
-    "main": "src/index",
-    "scripts": {
-        "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
-        "test:unit": "jest",
-        "type-check": "tsc --build"
-    },
-    "dependencies": {
-        "@mobily/ts-belt": "^3.13.1",
-        "@reduxjs/toolkit": "1.8.3",
-        "@suite-common/fiat-services": "workspace:*",
-        "@suite-common/redux-utils": "workspace:*",
-        "@suite-common/suite-config": "workspace:*",
-        "@suite-common/suite-utils": "workspace:*",
-        "@suite-common/test-utils": "workspace:*",
-        "@suite-common/wallet-config": "workspace:*",
-        "@suite-common/wallet-constants": "workspace:*",
-        "@suite-common/wallet-core": "workspace:*",
-        "@suite-common/wallet-types": "workspace:*",
-        "@suite-common/wallet-utils": "workspace:*",
-        "@trezor/blockchain-link": "workspace:*",
-        "@trezor/connect": "workspace:9.0.4",
-        "bignumber.js": "^9.1.0",
-        "date-fns": "^2.29.3"
-    },
-    "devDependencies": {
-        "jest": "^26.6.3"
-    }
-}
