--- suite-common/wallet-graph/src/__fixtures__/accountBalanceHistory.ts
@@ -1,662 +1,0 @@
-export const accountBalanceHistoryUsdRates = [
-    {
-        txs: 2,
-        received: '0.001',
-        sent: '0.001',
-        sentToSelf: '0',
-        time: 1421190000,
-        rates: {
-            usd: 172.1479,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 1,
-        received: '0.0155275',
-        sent: '0',
-        sentToSelf: '0',
-        time: 1422486000,
-        rates: {
-            usd: 232.5329422304941,
-        },
-        balance: '0.0155275',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 1,
-        received: '0',
-        sent: '0.0155275',
-        sentToSelf: '0',
-        time: 1422831600,
-        rates: {
-            usd: 240.0193,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 6,
-        received: '0.00444633',
-        sent: '0.00444633',
-        sentToSelf: '424633',
-        time: 1423177200,
-        rates: {
-            usd: 222.4872,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 1,
-        received: '0.00174998',
-        sent: '0',
-        sentToSelf: '0',
-        time: 1423522800,
-        rates: {
-            usd: 220.6493,
-        },
-        balance: '0.00174998',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 1,
-        received: '0',
-        sent: '0.000113',
-        sentToSelf: '163698',
-        time: 1423609200,
-        rates: {
-            usd: 219.0956,
-        },
-        balance: '0.00163698',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 1,
-        received: '0',
-        sent: '0.0002',
-        sentToSelf: '30000',
-        time: 1424214000,
-        rates: {
-            usd: 234.6238,
-        },
-        balance: '0.00143698',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 2,
-        received: '0.00113698',
-        sent: '0.00143698',
-        sentToSelf: '0',
-        time: 1428012000,
-        rates: {
-            usd: 253.6226999999999,
-        },
-        balance: '0.00113698',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 1,
-        received: '0',
-        sent: '0.00113698',
-        sentToSelf: '0',
-        time: 1428876000,
-        rates: {
-            usd: 223.9464,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 1,
-        received: '0.00093698',
-        sent: '0',
-        sentToSelf: '0',
-        time: 1433023200,
-        rates: {
-            usd: 229.6785,
-        },
-        balance: '0.00093698',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 1,
-        received: '0',
-        sent: '0.0003',
-        sentToSelf: '63698',
-        time: 1433714400,
-        rates: {
-            usd: 228.7013000000001,
-        },
-        balance: '0.00063698',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 1,
-        received: '0',
-        sent: '0.00063698',
-        sentToSelf: '0',
-        time: 1433800800,
-        rates: {
-            usd: 229.5263999999999,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 2,
-        received: '0.00043698',
-        sent: '0.00043698',
-        sentToSelf: '0',
-        time: 1434060000,
-        rates: {
-            usd: 230.5181,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 9,
-        received: '0.14895862',
-        sent: '0.14875862',
-        sentToSelf: '60000',
-        time: 1436392800,
-        rates: {
-            usd: 268.8928000000001,
-        },
-        balance: '0.0002',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 2,
-        received: '0.0001',
-        sent: '0.0002',
-        sentToSelf: '0',
-        time: 1438120800,
-        rates: {
-            usd: 288.7194990885924,
-        },
-        balance: '0.0001',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 2,
-        received: '0.00102316',
-        sent: '0.00112316',
-        sentToSelf: '0',
-        time: 1438466400,
-        rates: {
-            usd: 281.7267,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 16,
-        received: '0.002',
-        sent: '0.002',
-        sentToSelf: '0',
-        time: 1438639200,
-        rates: {
-            usd: 284.2393,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 2,
-        received: '0.0002',
-        sent: '0',
-        sentToSelf: '0',
-        time: 1443045600,
-        rates: {
-            usd: 233.9255,
-        },
-        balance: '0.0002',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 1,
-        received: '0',
-        sent: '0.0002',
-        sentToSelf: '0',
-        time: 1443477600,
-        rates: {
-            usd: 237.2262,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 2,
-        received: '0.01931681',
-        sent: '0.01931681',
-        sentToSelf: '0',
-        time: 1446073200,
-        rates: {
-            usd: 313.9073,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 5,
-        received: '0.001',
-        sent: '0.001',
-        sentToSelf: '219840',
-        time: 1472508000,
-        rates: {
-            usd: 576.08,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 2,
-        received: '0.0383',
-        sent: '0',
-        sentToSelf: '0',
-        time: 1473112800,
-        rates: {
-            usd: 610.6837999999999,
-        },
-        balance: '0.0383',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 2,
-        received: '0',
-        sent: '0.03812',
-        sentToSelf: '8000',
-        time: 1473199200,
-        rates: {
-            usd: 614.3264885962195,
-        },
-        balance: '0.00018',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 1,
-        received: '0',
-        sent: '0.00018',
-        sentToSelf: '0',
-        time: 1473458400,
-        rates: {
-            usd: 623.0588000000001,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 6,
-        received: '0.0016336',
-        sent: '0.0016336',
-        sentToSelf: '55040',
-        time: 1477605600,
-        rates: {
-            usd: 687.5587999999999,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 3,
-        received: '0.00114922',
-        sent: '0.00114922',
-        sentToSelf: '102632',
-        time: 1477692000,
-        rates: {
-            usd: 715.4103243232163,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 2,
-        received: '0.0015',
-        sent: '0.0015',
-        sentToSelf: '0',
-        time: 1477954800,
-        rates: {
-            usd: 729.2049999999999,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 1,
-        received: '0.00016708',
-        sent: '0',
-        sentToSelf: '0',
-        time: 1487977200,
-        rates: {
-            usd: 1149.4036781698383,
-        },
-        balance: '0.00016708',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 3,
-        received: '0.0033579',
-        sent: '0.00352498',
-        sentToSelf: '307138',
-        time: 1488063600,
-        rates: {
-            usd: 1180.6730000000002,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 4,
-        received: '0.00702825',
-        sent: '0',
-        sentToSelf: '0',
-        time: 1489705200,
-        rates: {
-            usd: 1074.0938695062455,
-        },
-        balance: '0.00702825',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 1,
-        received: '0',
-        sent: '0.00702825',
-        sentToSelf: '0',
-        time: 1489791600,
-        rates: {
-            usd: 961.9620000000001,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 10,
-        received: '0.02503907',
-        sent: '0.02503907',
-        sentToSelf: '2025425',
-        time: 1493157600,
-        rates: {
-            usd: 1287.32,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 2,
-        received: '0.00023714',
-        sent: '0.00023714',
-        sentToSelf: '0',
-        time: 1504130400,
-        rates: {
-            usd: 4702.93816308458,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 2,
-        received: '0.00009057',
-        sent: '0.00009057',
-        sentToSelf: '0',
-        time: 1508277600,
-        rates: {
-            usd: 5592.78222421231,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 5,
-        received: '0.001',
-        sent: '0.001',
-        sentToSelf: '46766',
-        time: 1511478000,
-        rates: {
-            usd: 8278.12843633886,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 2,
-        received: '0.0008517',
-        sent: '0.0008517',
-        sentToSelf: '0',
-        time: 1516057200,
-        rates: {
-            usd: 11723.9135206658,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-    {
-        txs: 2,
-        received: '0.0001',
-        sent: '0.0001',
-        sentToSelf: '0',
-        time: 1630706400,
-        rates: {
-            usd: 50015,
-        },
-        balance: '0',
-        descriptor: 'zpubabcd',
-    },
-];
-
-export const balanceMovementsBeforeStart1Hour = {
-    success: true,
-    payload: [
-        {
-            time: 1665619200,
-            txs: 1,
-            received: '521229',
-            sent: '0',
-            sentToSelf: '0',
-            rates: {
-                aed: 70353,
-                ars: 2892452,
-                aud: 30518,
-                bch: 171.145,
-                bdt: 1947545,
-                bhd: 7228.65,
-                bits: 1000087,
-                bmd: 19154.08,
-                bnb: 70.693,
-                brl: 101383,
-                btc: 1,
-                cad: 26467,
-                chf: 19102.65,
-                clp: 18091413,
-                cny: 137432,
-                czk: 484514,
-                dkk: 146759,
-                dot: 3104,
-                eos: 18431,
-                eth: 14.79821,
-                eur: 19729.41,
-                gbp: 17251.31,
-                hkd: 150352,
-                huf: 8525970,
-                idr: 293414667,
-                ils: 68445,
-                inr: 1574470,
-                jpy: 2811159,
-                krw: 27292650,
-                kwd: 5946.31,
-                link: 2680,
-                lkr: 7020801,
-                ltc: 367.071,
-                mmk: 40228308,
-                mxn: 382839,
-                myr: 89689,
-                ngn: 8405809,
-                nok: 206278,
-                nzd: 34139,
-                php: 1128558,
-                pkr: 4172630,
-                pln: 95588,
-                rub: 1220594,
-                sar: 71978,
-                sats: 100008707,
-                sek: 217068,
-                sgd: 27483,
-                thb: 726260,
-                try: 355395,
-                twd: 608887,
-                uah: 707519,
-                usd: 19154.08,
-                vef: 1917.9,
-                vnd: 458890408,
-                xag: 1004.66,
-                xau: 11.44,
-                xdr: 13633.45,
-                xlm: 164431,
-                xrp: 39245,
-                yfi: 2.430841,
-                zar: 350453,
-            },
-        },
-        {
-            time: 1665964800,
-            txs: 1,
-            received: '521088',
-            sent: '521229',
-            sentToSelf: '521088',
-            rates: {
-                aed: 70795,
-                ars: 2921372,
-                aud: 30970,
-                bch: 174.802,
-                bdt: 1968786,
-                bhd: 7267.9,
-                bits: 1000181,
-                bmd: 19274.21,
-                bnb: 70.754,
-                brl: 102710,
-                btc: 1,
-                cad: 26698,
-                chf: 19341.29,
-                clp: 18541988,
-                cny: 138601,
-                czk: 486633,
-                dkk: 147156,
-                dot: 3115,
-                eos: 18309,
-                eth: 14.74736,
-                eur: 19784.54,
-                gbp: 17137.15,
-                hkd: 151294,
-                huf: 8253318,
-                idr: 297346596,
-                ils: 68414,
-                inr: 1588447,
-                jpy: 2864360,
-                krw: 27788733,
-                kwd: 5983.91,
-                link: 2686,
-                lkr: 7038200,
-                ltc: 374.344,
-                mmk: 40493753,
-                mxn: 386017,
-                myr: 90627,
-                ngn: 8387753,
-                nok: 205198,
-                nzd: 34564,
-                php: 1137179,
-                pkr: 4217784,
-                pln: 95463,
-                rub: 1207530,
-                sar: 72429,
-                sats: 100018061,
-                sek: 217322,
-                sgd: 27475,
-                thb: 737239,
-                try: 358305,
-                twd: 616332,
-                uah: 712110,
-                usd: 19274.21,
-                vef: 1929.93,
-                vnd: 465215670,
-                xag: 1048.54,
-                xau: 11.7,
-                xdr: 13718.96,
-                xlm: 169782,
-                xrp: 40433,
-                yfi: 2.485325,
-                zar: 352317,
-            },
-        },
-        {
-            time: 1666656000,
-            txs: 1,
-            received: '464814',
-            sent: '516088',
-            sentToSelf: '464814',
-            rates: {
-                aed: 70967,
-                ars: 2988151,
-                aud: 30512,
-                bch: 179.236,
-                bdt: 1955580,
-                bhd: 7284.66,
-                bits: 999622,
-                bmd: 19321.36,
-                bnb: 70.591,
-                brl: 102552,
-                btc: 1,
-                cad: 26456,
-                chf: 19309.05,
-                clp: 19043777,
-                cny: 140331,
-                czk: 478167,
-                dkk: 145345,
-                dot: 3270,
-                eos: 18128,
-                eth: 14.380602,
-                eur: 19538.69,
-                gbp: 17081.49,
-                hkd: 151669,
-                huf: 8094645,
-                idr: 301374575,
-                ils: 68708,
-                inr: 1599070,
-                jpy: 2875965,
-                krw: 27918613,
-                kwd: 5996.07,
-                link: 2798,
-                lkr: 7056569,
-                ltc: 367.844,
-                mmk: 40378145,
-                mxn: 384939,
-                myr: 91545,
-                ngn: 8470524,
-                nok: 202821,
-                nzd: 33857,
-                php: 1136763,
-                pkr: 4241772,
-                pln: 93518,
-                rub: 1197922,
-                sar: 72607,
-                sats: 99962225,
-                sek: 215346,
-                sgd: 27456,
-                thb: 738462,
-                try: 359588,
-                twd: 624588,
-                uah: 710107,
-                usd: 19321.36,
-                vef: 1934.65,
-                vnd: 480229287,
-                xag: 1000.28,
-                xau: 11.7,
-                xdr: 13752.52,
-                xlm: 175075,
-                xrp: 42631,
-                yfi: 2.536798,
-                zar: 355711,
-            },
-        },
-    ],
-};
