--- suite-common/wallet-graph/src/constants.ts
@@ -1,6 +1,0 @@
-export const networkAccountsNotFoundError = 'NETWORK_ACCOUNTS_NOT_FOUND';
-export const accountNotFoundError = 'ACCOUNT_NOT_FOUND';
-export const actionPrefix = '@common/wallet-graph';
-
-// more than 190 points will cause graph not to show, also it's better to show less because of performance
-export const MAX_GRAPH_POINTS_NUMBER = 80;
