--- suite-common/wallet-graph/src/index.ts
@@ -1,5 +1,0 @@
-export * from './graphReducer';
-export * from './graphThunks';
-export * from './config';
-export * from './graphUtils';
-export * from './types';
