--- suite-common/wallet-graph/tsconfig.json
@@ -1,20 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": { "outDir": "libDev" },
-    "references": [
-        { "path": "../fiat-services" },
-        { "path": "../redux-utils" },
-        { "path": "../suite-config" },
-        { "path": "../suite-utils" },
-        { "path": "../test-utils" },
-        { "path": "../wallet-config" },
-        { "path": "../wallet-constants" },
-        { "path": "../wallet-core" },
-        { "path": "../wallet-types" },
-        { "path": "../wallet-utils" },
-        {
-            "path": "../../packages/blockchain-link"
-        },
-        { "path": "../../packages/connect" }
-    ]
-}
