+++ suite-native/graph/package.json
@@ -7,24 +7,26 @@
     "main": "src/index",
     "scripts": {
         "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
+        "test:unit": "jest -c ../../jest.config.base.js",
         "type-check": "tsc --build"
     },
     "dependencies": {
+        "@mobily/ts-belt": "^3.13.1",
         "@shopify/react-native-skia": "0.1.162",
         "@suite-common/formatters": "workspace:*",
+        "@suite-common/graph-neue": "workspace:*",
         "@suite-common/wallet-types": "workspace:*",
         "@suite-native/atoms": "workspace:*",
         "@suite-native/test-utils": "workspace:*",
         "@trezor/icons": "workspace:*",
         "@trezor/styles": "workspace:*",
+        "date-fns": "^2.29.1",
         "react": "18.2.0",
         "react-native": "0.70.5",
         "react-native-gesture-handler": "^2.7.0",
         "react-native-graph": "^0.3.0",
         "react-native-reanimated": "^2.12.0",
-        "react-redux": "7.2.2",
-        "@suite-common/graph-neue": "workspace:*",
-        "date-fns": "^2.29.1"
+        "react-redux": "7.2.2"
     },
     "devDependencies": {
         "jest": "^26.6.3",
