+++ suite-native/module-accounts/package.json
@@ -16,10 +16,10 @@
         "@react-navigation/stack": "^6.2.2",
         "@reduxjs/toolkit": "1.8.3",
         "@suite-common/formatters": "workspace:*",
+        "@suite-common/graph-neue": "workspace:*",
         "@suite-common/suite-types": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
         "@suite-common/wallet-core": "workspace:*",
-        "@suite-common/wallet-graph": "workspace:*",
         "@suite-common/wallet-types": "workspace:*",
         "@suite-common/wallet-utils": "workspace:*",
         "@suite-native/accounts": "workspace:*",
