+++ suite-native/module-accounts/src/components/AccountBalance.tsx
@@ -8,8 +8,7 @@ import { Box, DiscreetText, Divider } from '@suite-native/atoms';
 import { CryptoIcon } from '@trezor/icons';
 import { AccountsRootState, selectAccountByKey } from '@suite-common/wallet-core';
 import { useFormatters } from '@suite-common/formatters';
-import { formatNetworkAmount } from '@suite-common/wallet-utils';
-import { EnhancedGraphPointWithCryptoBalance } from '@suite-native/graph';
+import { emptyGraphPoint, EnhancedGraphPointWithCryptoBalance } from '@suite-native/graph';
 
 type AccountBalanceProps = {
     accountKey: string;
@@ -19,17 +18,10 @@ const cryptoIconStyle = prepareNativeStyle(utils => ({
     marginRight: utils.spacings.small / 2,
 }));
 
-const emptyPoint: EnhancedGraphPointWithCryptoBalance = {
-    value: 0,
-    date: new Date(),
-    timestamp: 0,
-    cryptoBalance: '0',
-};
-
-const selectedPointAtom = atom<EnhancedGraphPointWithCryptoBalance>(emptyPoint);
+const selectedPointAtom = atom<EnhancedGraphPointWithCryptoBalance>(emptyGraphPoint);
 
 // reference is usually first point, same as Revolut does in their app
-const referencePointAtom = atom<EnhancedGraphPointWithCryptoBalance>(emptyPoint);
+const referencePointAtom = atom<EnhancedGraphPointWithCryptoBalance>(emptyGraphPoint);
 
 export const writeOnlySelectedPointAtom = atom<null, EnhancedGraphPointWithCryptoBalance>(
     null, // it's a convention to pass `null` for the first argument
@@ -54,9 +46,6 @@ export const AccountBalance = ({ accountKey }: AccountBalanceProps) => {
 
     if (!account) return null;
 
-    // TODO this should be done with formatters once they're prepared
-    const cryptoAmount = formatNetworkAmount(account.availableBalance, account.symbol);
-
     return (
         <Box>
             <Box marginBottom="large" justifyContent="center" alignItems="center">
@@ -65,7 +54,7 @@ export const AccountBalance = ({ accountKey }: AccountBalanceProps) => {
                         <CryptoIcon name={account.symbol} />
                     </Box>
                     <DiscreetText color="gray600" typography="hint">
-                        {CryptoAmountFormatter.format(cryptoAmount, {
+                        {CryptoAmountFormatter.format(selectedPoint.cryptoBalance, {
                             symbol: account.symbol,
                         })}
                     </DiscreetText>
