+++ suite-native/module-home/tsconfig.json
@@ -9,6 +9,9 @@
             "path": "../../suite-common/formatters"
         },
         {
+            "path": "../../suite-common/graph-neue"
+        },
+        {
             "path": "../../suite-common/suite-utils"
         },
         {
@@ -16,9 +19,6 @@
         },
         {
             "path": "../../suite-common/wallet-core"
-        },
-        {
-            "path": "../../suite-common/wallet-graph"
         },
         {
             "path": "../../suite-common/wallet-types"
